+++ CHANGELOG.md
@@ -2,6 +2,11 @@ yum-fedora Cookbook CHANGELOG
 ======================
 This file is used to list changes made in each version of the yum-oracle cookbook.
 
+v1.0.4 (2020-11-13)
+-------------------
+- support Oracle Enterprise Linux 8
+- cookstyle cookbook
+
 v1.0.3 (2016-03-01)
 -------------------
 - added in support for `make_cache` attribute